chore: Split up Add and Edit modals
