+++ apps/editor.planx.uk/src/pages/FlowEditor/components/Team/types.ts
@@ -31,7 +31,7 @@ export interface MembersTableProps {
   showRemoveMemberButton?: boolean;
 }
 
-export interface AddNewEditorFormValues {
+export interface UserFormValues {
   email: string;
   firstName: string;
   lastName: string;
@@ -50,3 +50,5 @@ type SharedModalProps = {
 };
 
 export type EditorModalProps = SharedModalProps & ModalState;
+
+export const DEMO_TEAM_ID = 32;
